+++ src/@lekoarts/gatsby-theme-emma/gatsby-plugin-theme-ui/styles/buttons.js
@@ -1,14 +1,13 @@
 export const buttons = {
 	toggle: {
-		color: 'white',
+		color: 'gs800',
 		border: 'none',
 		backgroundColor: 'primary',
 		cursor: 'pointer',
 		alignSelf: 'center',
-		px: 3,
-		py: 2,
-		ml: 3,
+		//px: 3,
+		//py: 2,
+		//ml: 3,
 		fontFamily: 'body',
-		boderRadius: 5,
 	},
 };