+++ src/templates/home.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable react/no-unescaped-entities */
+
 import * as React from 'react';
 // Providers!
 import { MDXProvider } from '@mdx-js/react';
@@ -6,24 +8,183 @@ import useNavigation from '@lekoarts/gatsby-theme-emma/src/hooks/use-navigation'
 import useSiteMetadata from '@lekoarts/gatsby-theme-emma/src/hooks/use-site-metadata';
 
 // Third Party Components!
-import { Image, Card as CardUI, Flex } from 'theme-ui';
+
+import { Flex, Box } from 'theme-ui';
 // Custom Components!
-import ProjectCard from 'components/project-card.tsx';
-import Card, { CardVertical } from 'components/cards/card';
+import SkillCard from 'components/cards/skill-card';
+import CardContact from 'components/cards/contact-card';
+import ExperienceCard from 'components/cards/experience-card';
+import ProjectCard from 'components/cards/project-card';
 import Seo from '@lekoarts/gatsby-theme-emma/src/components/seo';
 import Footer from '@lekoarts/gatsby-theme-emma/components/footer';
 import Header from '@lekoarts/gatsby-theme-emma/src/components/header';
-
+import Hero from 'components/hero';
 // Sections
-import Hero from '../../content/sections/hero';
 import Skills from '../../content/sections/skills';
 import Contact from '../../content/sections/contact';
 import Projects from '../../content/sections/projects';
 import Experiences from '../../content/sections/experiences';
 
+import HeroContent from '../../content/sections/hero';
+
+import SectionLinks from 'components/links/section-links';
+
+
 const components = {
-	ProjectCard,
+	CardContact,
 };
+
+const FlexContent = ({ children }) => (
+	<Flex
+		sx={{
+			gap: '1.5rem',
+			textAlign: 'center',
+			alignItems: 'center',
+			flexDirection: 'column',
+		}}
+	>
+		{children}
+	</Flex>
+);
+const HeroSection = () => (
+	<Hero>
+		<HeroContent />
+	</Hero>
+);
+const ProjectsSection = () => (
+	<Box as="section">
+		<h2>Projects</h2>
+		<FlexContent>
+			<Projects />
+			<ProjectCard
+				title="Gaboland"
+				borderColor="success"
+				to="https://github.com/gabrielba15/portfolio"
+			>
+				My personal website or portfolio. This projects is built with Gatsby,
+				React and Theme UI.
+			</ProjectCard>
+			<ProjectCard
+				title="Giffy"
+				to="https://github.com/gabrielba15/react-giffy"
+			>
+				Gifs App using React, Styled Components and consuming giphy api.
+			</ProjectCard>
+			<ProjectCard
+				title="PHP & MYSQL Parcial"
+				borderColor="success"
+				to="https://github.com/gabrielba15/php_mysql_parcial"
+			>
+				CRUD App using PHP, MySQL and Bootstrap for a university assessment.
+			</ProjectCard>
+		</FlexContent>
+		<SectionLinks
+			icon="github"
+			text="All Projects"
+			to={{ link: 'https://github.com/gaboland', button: '/projects' }}
+		/>
+	</Box>
+);
+
+const SkillsSection = () => (
+	<Box as="section">
+		<h2>Skills</h2>
+		<FlexContent>
+			<Skills />
+			<SkillCard
+				title="Programming"
+				borderColor="success"
+				to="/skills/programming"
+			>
+				My life as a programmer started in 2019. I have 3 years of experiences
+				and I'm still learning.
+			</SkillCard>
+			<SkillCard title="Design" to="/skills/design">
+				I have 2 year designing and learning technologies associeted with it and
+				I still don't stop.
+			</SkillCard>
+			<SkillCard
+				title="Autodidact"
+				borderColor="success"
+				to="/skills/autodidact"
+			>
+				since 2017 I haven't stopped learning somenthing new every single day.
+			</SkillCard>
+		</FlexContent>
+		<SectionLinks
+			icon="linkedin"
+			text="All Skills"
+			to={{ link: 'https://linkedin/in/gabrielbencomo', button: '/skills' }}
+		/>
+	</Box>
+);
+const ExperiencesSection = () => (
+	<Box as="section">
+		<h2>Experiences</h2>
+		<FlexContent>
+			<Experiences />
+			<ExperienceCard
+				title="Designer"
+				company="Freelance"
+				location="Barinas, Venezuela"
+				year="1 year"
+			/>
+			<ExperienceCard
+				title="Web Developer"
+				company="Freelance"
+				location="Barinas, Venezuela"
+				year="2 years"
+				borderColor="success"
+			/>
+			<ExperienceCard
+				title="UI Designer"
+				company="Freelance"
+				location="Barinas, Venezuela"
+				year="1 year"
+			/>
+		</FlexContent>
+		<SectionLinks
+			icon="linkedin"
+			text="All Experiences"
+			to={{
+				link: 'https://linkedin/in/gabrielbencomo',
+				button: '/experiences',
+			}}
+		/>
+	</Box>
+);
+const ContactSection = () => (
+	<Box as="section">
+		<h2>Contact</h2>
+		<FlexContent>
+			<Contact />
+			<CardContact
+				title="Work Contact"
+				icons={{ first: 'email', second: 'linkedin' }}
+				to={{
+					first: 'https://linkedin.com/in/gabrielbencomo',
+					second: 'https://linkedin.com/in/gabrielbencomo',
+				}}
+			>
+				If you are an entrepreneur or somebody who are looking for someone who
+				develop your app or website, you can contact me from 8:00h to 11:00h and
+				from 15:00h to 18:h ET. for an immediate response.
+			</CardContact>
+			<CardContact
+				title="Friendly Contact"
+				icons={{ first: 'telegram', second: 'twitter' }}
+				to={{
+					first: 'https://t.me/gabrielba15',
+					second: 'https://twitter.com/gabrielbca15',
+				}}
+			>
+				If you're an old friend or a stranger who wants to contact me for
+				anything (say hello, connection, etc.) write at any time through the
+				media but do not expect an immediate response.
+			</CardContact>
+		</FlexContent>
+	</Box>
+);
 
 const Home = () => {
 	const meta = useSiteMetadata();
@@ -32,119 +193,22 @@ const Home = () => {
 		<MDXProvider components={components}>
 			<Seo />
 			<Header meta={meta} nav={nav} />
-			<main className="container">
-				<CardUI
-					sx={{
-						width: 200,
-						mx: 'auto',
-						pt: 60,
-					}}
-				>
-					<Image
-						as={Image}
-						sx={{
-							borderRadius: '50%',
-							width: 200,
-							height: 200,
-						}}
-						src="me.jpg"
-						alt="Selfie Of Me"
-					/>
-				</CardUI>
-				<Flex
-					sx={{
-						flexDirection: 'column',
-						textAlign: 'center',
-						mb: 60,
-					}}
-				>
-					<Hero />
-				</Flex>
-				<h2>Projects</h2>
-				<Flex
-					sx={{
-						flexDirection: 'column',
-						textAlign: 'center',
-						alignItems: 'center',
-					}}
-				>
-					<Projects />
-					<CardVertical title="Gaboland" borderColor="success">
-						My personal website or portfolio. This projects is built with
-						Gatsby, React and Theme UI.
-					</CardVertical>
-					<CardVertical title="Giffy">
-						Gifs App using React, Styled Components and consuming giphy api.
-					</CardVertical>
-					<CardVertical title="PHP & MYSQL Parcial" borderColor="success">
-						CRUD App using PHP, MySQL and Bootstrap for a university assessment.
-					</CardVertical>
-				</Flex>
-				<h2>Skills</h2>
-				<Flex
-					sx={{
-						flexDirection: 'column',
-						textAlign: 'center',
-						alignItems: 'center',
-					}}
-				>
-					<Skills />
-					<CardVertical title="Programming">
-						My life as a programmer started in 2019. I have 3 years of
-						experiences and I’m still learning.
-					</CardVertical>
-					<CardVertical title="Design" borderColor="success">
-						I have 2 year designing and learning technologies associeted with it
-						and I still don’t stop.
-					</CardVertical>
-					<CardVertical title="Autodidact">
-						since 2017 I haven’t stopped learning somenthing new every single
-						day.
-					</CardVertical>
-				</Flex>
-				<h2>Experiences</h2>
-				<Flex
-					sx={{
-						flexDirection: 'column',
-						textAlign: 'center',
-						alignItems: 'center',
-					}}
-				>
-					<Experiences />
-					<CardVertical title="Gaboland">
-						My personal website or portfolio. This projects is built with
-						Gatsby, React and Theme UI.
-					</CardVertical>
-					<CardVertical title="Gaboland">
-						My personal website or portfolio. This projects is built with
-						Gatsby, React and Theme UI.
-					</CardVertical>
-					<CardVertical title="Gaboland">
-						My personal website or portfolio. This projects is built with
-						Gatsby, React and Theme UI.
-					</CardVertical>
-				</Flex>
-				<h2>Contact</h2>
-				<Flex
-					sx={{
-						flexDirection: 'column',
-						textAlign: 'center',
-						alignItems: 'center',
-					}}
-				>
-					<Contact />
-					<Card title="Work Contact">
-						If you are an entrepreneur or somebody who are looking for someone
-						who develop your app or website, you can contact me from 8:00h to
-						11:00h and from 15:00h to 18:h ET. for an immediate response.
-					</Card>
-					<Card title="Friendly Contact">
-						If you’re an old friend or a stranger who wants to contact me for
-						anything (say hello, connection, etc.) write at any tiime through
-						the media but do not expect an immediate response.
-					</Card>
-				</Flex>
-			</main>
+			<Box
+				as="main"
+				sx={{
+					display: 'grid',
+					gridGap: '3rem',
+					maxWidth: 720,
+					margin: '0 auto',
+					padding: '0 24px',
+				}}
+			>
+				<HeroSection />
+				<ProjectsSection />
+				<SkillsSection />
+				<ExperiencesSection />
+				<ContactSection />
+			</Box>
 			<Footer />
 		</MDXProvider>
 	);