+++ docs/CHANGELOG.md
@@ -1,53 +1,157 @@
 # Change Log
 
-### 0.5.3 (Jan 14, 2022)
+
+## 🛸 Minor [0.6.0](https://github.com/gabrielba15/portfolio/releases/tag/v0.6.0) (Jan 28, 2022)
+
+##### 🔥 Features
+
+- NEW: Page -> Home Page v1.0-beta ([e42c5db3c](https://github.com/gabrielba15/portfolio/commit/e42c5db3c4ae633b8b4e13860b976739658c204e)).
+
+##### 🐞 Bugs 
+
+- FIXED: Theme -> Gradient Color Disaster ([e6e6baa61](https://github.com/gabrielba15/portfolio/commit/e6e6baa61732b8f5e02861157e12b1153295e5ec)).
+
+
+### 0.5.1 (Jan 14, 2022)
+
+
+### 🧪 Patch [0.5.8](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.8) (Jan 27, 2022) 
+
+##### 🔥 Features
+
+- NEW: Component -> Link With Icon v1.0 ([e661cef83](https://github.com/gabrielba15/portfolio/commit/e661cef836914fd55dfb1dd641289bba1d2c61bd)).
+- NEW: Component -> Section Links v1.0 ([e661cef83](https://github.com/gabrielba15/portfolio/commit/e661cef836914fd55dfb1dd641289bba1d2c61bd)).
+
+##### 🔄 Refactor
+
+- CHANGES: Component → Button With Icon [Padding Right from 10 to 0.] ([1329a3e76](https://github.com/gabrielba15/portfolio/commit/1329a3e7605e803a77dec314d98daaf866a068a9)).
+- CHANGES: Component -> Hero [Letter Spacing Added] ([e8b9f166f](https://github.com/gabrielba15/portfolio/commit/e8b9f166f3a3b95af3cfec34cf33d3a05569fd0d)).
+
+### 🧪 Patch [0.5.7](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.7) (Jan 25, 2022) 
+
+##### 🔥 Features
+
+- NEW: Component -> Project Card v1.0 ([d44139381](https://github.com/gabrielba15/portfolio/commit/d44139381dd21d479737d89ac6537e9612eb2f2b)).
+
+- NEW: Component -> Hero v0.1 ([929216ee1](https://github.com/gabrielba15/portfolio/commit/929216ee1e79b6cb9369d09b14b5b20c874dc30c)).
+
+##### 🔄 Refactor
+
+- CHANGES: Page -> Home [[Hero Component Added To UI](https://github.com/gabrielba15/portfolio/commit/929216ee1e79b6cb9369d09b14b5b20c874dc30c)] ([88d43c3a7](https://github.com/gabrielba15/portfolio/commit/88d43c3a70028111efd2acac755852fe6c36d8d4)).
+
+- CHANGES: Component -> Button With Icon [Border Color Changed] ([88d43c3a7](https://github.com/gabrielba15/portfolio/commit/88d43c3a70028111efd2acac755852fe6c36d8d4)).
+
+- CHANGES: Component -> Footer [Icons Added, Button With Icon Added] ([88d43c3a7](https://github.com/gabrielba15/portfolio/commit/88d43c3a70028111efd2acac755852fe6c36d8d4)).
+
+- CHANGES: Theme -> Button Styles Changed [Border Color Deleted] ([88d43c3a7](https://github.com/gabrielba15/portfolio/commit/**88d43c3a70028111efd2acac755852fe6c36d8d4**)).
+
+##### 🐞 Bugs 
+- FIXED: Component -> Loop Skill Card ([1a6d5de25](https://github.com/gabrielba15/portfolio/commit/1a6d5de255ad25905211f1e8c978287dab52ff0f)).
+
+##### 📦 Packages 
+
+- UPDATED: Bump -> nanoid ([c190ad0ee](https://github.com/gabrielba15/portfolio/commit/c190ad0eefdf9130a45b0ddc507cf3785973ea4e))
+### 🧪 Patch [0.5.6](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.6) (Jan 23, 2022)
+
+##### 🔥 Features
+
+- NEW: Component -> Skill Card v1.0 ([118490ea5](https://github.com/gabrielba15/portfolio/commit/118490ea582c1941ffc3468bb448d5c79a6c1eb1)).
+
+##### 🔄 Refactor
+
+- CHANGES: Page -> Home [[Skill Card Component Added To UI](https://github.com/gabrielba15/portfolio/commit/118490ea582c1941ffc3468bb448d5c79a6c1eb1)] ([7b4443dd7](https://github.com/gabrielba15/portfolio/commit/7b4443dd727acfb4892ea5ba78b5ade29b356b36)).
+
+- CHANGES: Component -> Button With Icon [Side Prop, New Default Styles]  ([8a446cbc4](https://github.com/gabrielba15/portfolio/commit/8a446cbc4e6a53888a07acb90efaca0b0afa6a1f)).
+
+- CHANGES: Component -> Contact Card [Capitalizer Polyfill]  ([8a446cbc4](https://github.com/gabrielba15/portfolio/commit/8a446cbc4e6a53888a07acb90efaca0b0afa6a1f)).
+
+
+### 🧪 Patch [0.5.5](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.5) (Jan 21, 2022)
+
+##### 🔥 Features
+
+- NEW: Component -> Experience Card v1.0 ([5ba784b02](https://github.com/gaboland/portfolio/commit/5ba784b028bc30b11d5651b43042ef4dced08f58)).
+- NEW: Component -> Small Card v1.0 ([ceb12d7e7](https://github.com/gabrielba15/portfolio/commit/ceb12d7e7dd9ae12b8dedf480bfe7b9d58c523c8)).
+- NEW: Component -> Badge v1.0 ([a5c731cd3](https://github.com/gabrielba15/portfolio/commit/a5c731cd3ad0f2912cf001d55670bb9ae430a4c7)).
+
+##### 🔄 Refactor
+
+- CHANGES: Page -> Home [[Experience Card Component Added](https://github.com/gaboland/portfolio/commit/5ba784b028bc30b11d5651b43042ef4dced08f58)] ([d4e83423f](https://github.com/gaboland/portfolio/commit/d4e83423fbae5d6c82f63959d42c65279454bc84)).
+
+### 🧪 Patch [0.5.4](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.4) (Jan 20, 2022)
+
+##### 🔥 Features
+
+- NEW: Icons - Bootstrap ([c65e05160](https://github.com/gaboland/portfolio/commit/c65e051605d454d68f68abab1b215046e78164a8)).
+- NEW: Component -> Contact Card v1.0 ([ce7f68e94](https://github.com/gaboland/portfolio/commit/ce7f68e94acad77cb4f3bd51b1192b535dcb51f2)).
+- NEW: Component -> Button With Icon v1.0 ([9429128d1](https://github.com/gaboland/portfolio/commit/9429128d1daa583635f548c3241787e38722dd29)).
+
+##### 🔄 Refactor
+
+- CHANGES: Page -> Home [[Contact Card Component Added](https://github.com/gaboland/portfolio/commit/ce7f68e94acad77cb4f3bd51b1192b535dcb51f2)] ([9b81e7f98](https://github.com/gaboland/portfolio/commit/9b81e7f98a3340f00e7d29aa4064999638133d18)).
+- CHANGES: Component -> Contact Card [[Prop borderColor fixed](https://github.com/gaboland/portfolio/commit/ce7f68e94acad77cb4f3bd51b1192b535dcb51f2)] ([0ed59d801](https://github.com/gaboland/portfolio/commit/0ed59d8014d7e1796ba72c7fd4f4c5ad0d2cd0d3)).
+
+##### 📦 Packages
+
+- UPDATED: engine io [[#5](https://github.com/gabrielba15/portfolio/pull/5)] ([783503a10](https://github.com/gaboland/portfolio/commit/783503a10b6b7551e13e6e3083ee3a5534f2ae40)).
+- UPDATED: follow-redirects [[#6](https://github.com/gabrielba15/portfolio/pull/6)] ([08038a942](https://github.com/gaboland/portfolio/commit/08038a942d74c9b3f04f58d0335b03d60613c1aa)).
+
+##### 🟢 Merges
+
+- MERGING: main -> development [[#7](https://github.com/gabrielba15/portfolio/pull/7)] ([26d15881f](https://github.com/gaboland/portfolio/commit/26d15881f182a652200903793090592c6d5081df)).
+
+### 🧪 Patch [0.5.3](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.3) (Jan 14, 2022)
 
 ### Bug
 
 - Vercel Deploy Fixed ([bad6a8546](https://github.com/gaboland/portfolio/commit/bad6a854671a286aff194ad087a57c15d948cdac)).
-### 0.5.2 (Jan 14, 2022)
+### 🧪 Patch [0.5.2](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.2) (Jan 14, 2022)
 
 ### Bug
 
 - Vercel Deploy ([acfc09404](https://github.com/gaboland/portfolio/commit/acfc09404737a1bbe5650127b45ec997f21f4e77)).
 
-### 0.5.1 (Jan 14, 2022)
+### 🧪 Patch [0.5.1](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.1) (Jan 14, 2022)
 
 ### Config
 
 - New Gatsby Config ([1056f4738](https://github.com/gaboland/portfolio/commit/1056f4738a266192a7fa258cf414dbf67dd4ce17)).
 - All Modules Updated to Latest Version ([1056f4738](https://github.com/gaboland/portfolio/commit/1056f4738a266192a7fa258cf414dbf67dd4ce17)).
 
-## 0.5.0 (Jan 13, 2022)
+## 🛸 Minor [0.5.0](https://github.com/gabrielba15/portfolio/releases/tag/v0.5.0) (Jan 13, 2022)
 
 - New Border Color to Cards ([d91eef087](https://github.com/gaboland/portfolio/commit/d91eef087dc8b7c26194c5ca868438da5fa9dc08)).
 - New Card Info ([d91eef087](https://github.com/gaboland/portfolio/commit/d91eef087dc8b7c26194c5ca868438da5fa9dc08)).
 
-## 0.4.0 (Jan 11, 2022)
+## 🛸 Minor [0.4.0](https://github.com/gabrielba15/portfolio/releases/tag/v0.4.0) (Jan 11, 2022)
 
 - Selfie ([942cfb9ac](https://github.com/gaboland/portfolio/commit/942cfb9ac3fe9a4ff3c75084be7c12bfe3dd766c)).
 - Content Centered ([942cfb9ac](https://github.com/gaboland/portfolio/commit/942cfb9ac3fe9a4ff3c75084be7c12bfe3dd766c)).
 
-### 0.3.2 (Jan 10, 2022)
+### 🧪 Patch [0.3.2](https://github.com/gabrielba15/portfolio/releases/tag/v0.3.2) (Jan 10, 2022)
 
 ### Bug
 
 - Vercel Deploy Fixed ([19440a41c](https://github.com/gaboland/portfolio/commit/19440a41c285e25b87f77197f3674d63b911afc6)).
 
-### 0.3.1 (Jan 10, 2022)
+### 🧪 Patch [0.3.1](https://github.com/gabrielba15/portfolio/releases/tag/v0.3.1) (Jan 10, 2022)
 
 ### Bug
 
 - Vercel Deploy ([6e104402c](https://github.com/gaboland/portfolio/commit/6e104402c58dee9b96443e6d093c8cc0d6e9bd33)).
 
-## 0.3.0 (Jan 10, 2022)
+
+## 🛸 Minor [0.3.0](https://github.com/gabrielba15/portfolio/releases/tag/v0.3.0) (Jan 10, 2022)
+
 
 - Vertical Card v.0.0.1 ([673f1147c](https://github.com/gaboland/portfolio/commit/673f1147cc51d5eb824a924255da60bef07645bc)).
 - Content Card v.0.0.1 ([673f1147c](https://github.com/gaboland/portfolio/commit/673f1147cc51d5eb824a924255da60bef07645bc)).
 
 #### Features
 
-### 0.2.1 (Jan 9, 2022)
+
+### 🧪 Patch [0.2.1](https://github.com/gabrielba15/portfolio/releases/tag/v0.2.1) (Jan 9, 2022)
+
 
 #### Features
 
@@ -61,37 +165,37 @@
 
 - Gradient Color Disaster ([e6d320240](https://github.com/gaboland/portfolio/commit/e6d320240849fd0b09a952a87eefc2b9bd5bab70)).
 
-## 0.2.0 (Jan 8, 2022)
+## 🛸 Minor [0.2.0](https://github.com/gabrielba15/portfolio/releases/tag/v0.2.0) (Jan 8, 2022)
 
-- Projects and Project Pages Added ([6fe2bbe56](https://github.com/gaboland/gaboland-cara-starter/commit/6fe2bbe56c93a36941d90eea6974b2d1dd5ffcdc)).
-- Hero, Projects, Skills, Experiences, Contact Sections Added to Main Page ([6fe2bbe56](https://github.com/gaboland/gaboland-cara-starter/commit/6fe2bbe56c93a36941d90eea6974b2d1dd5ffcdc)).
+- Projects and Project Pages Added ([6fe2bbe56](https://github.com/gabrielba15/portfolio/commit/6fe2bbe56c93a36941d90eea6974b2d1dd5ffcdc)).
+- Hero, Projects, Skills, Experiences, Contact Sections Added to Main Page ([6fe2bbe56](https://github.com/gabrielba15/portfolio/commit/6fe2bbe56c93a36941d90eea6974b2d1dd5ffcdc)).
 
-## 0.1.0 (Jan 7, 2022)
+## 🛸 Minor [0.1.0](https://github.com/gabrielba15/portfolio/releases/tag/v0.1.0) (Jan 7, 2022)
 
-- Footer ([6d34c0dc5](https://github.com/gaboland/gaboland-cara-starter/commit/6d34c0dc5be3707fe0c28a3e7c6638bd3ffe97fd)).
-- Flex Header ([6d34c0dc5](https://github.com/gaboland/gaboland-cara-starter/commit/6d34c0dc5be3707fe0c28a3e7c6638bd3ffe97fd)).
-- Theme Changer Button with new styles ([6d34c0dc5](https://github.com/gaboland/gaboland-cara-starter/commit/6d34c0dc5be3707fe0c28a3e7c6638bd3ffe97fd)).
+- Footer ([6d34c0dc5](https://github.com/gabrielba15/portfolio/commit/6d34c0dc5be3707fe0c28a3e7c6638bd3ffe97fd)).
+- Flex Header ([6d34c0dc5](https://github.com/gabrielba15/portfolio/commit/6d34c0dc5be3707fe0c28a3e7c6638bd3ffe97fd)).
+- Theme Changer Button with new styles ([6d34c0dc5](https://github.com/gabrielba15/portfolio/commit/6d34c0dc5be3707fe0c28a3e7c6638bd3ffe97fd)).
 
-### 0.0.4 (Jan 6, 2022)
-
-#### Features
-
-- Home Component ([73889c999](https://github.com/gaboland/gaboland-cara-starter/commit/73889c999c03e3af70c68a0dcf11e525dde7dd50)).
-
-### 0.0.3 (Jan 5, 2022)
+### 🧪 Patch [0.0.4](https://github.com/gabrielba15/portfolio/releases/tag/v0.0.4) (Jan 6, 2022)
 
 #### Features
 
-- Changer Button Component ([8ee1d1c1f](https://github.com/gaboland/gaboland-cara-starter/commit/8ee1d1c1ffac80e16f27c4075e6b0fe7db11c8a3)).
+- Home Component ([73889c999](https://github.com/gabrielba15/portfolio/commit/73889c999c03e3af70c68a0dcf11e525dde7dd50)).
 
-### 0.0.2 (Dec 30, 2021)
-
-#### Features
-
-- Alt Name Gaboland ([b0a8ebc0c](https://github.com/gaboland/gaboland-cara-starter/commit/b0a8ebc0cc38579c68b22a967ba814556af1b3aa)).
-
-### 0.0.1 (Dec 29, 2021)
+### 🧪 Patch [0.0.3](https://github.com/gabrielba15/portfolio/releases/tag/v0.0.3) (Jan 5, 2022)
 
 #### Features
 
-- Initial Commit ([173d18f8f](https://github.com/gaboland/gaboland-cara-starter/commit/173d18f8fbd7862a7e1ee63a9723fb5d2c438d29)).
+- Changer Button Component ([8ee1d1c1f](https://github.com/gabrielba15/portfolio/commit/8ee1d1c1ffac80e16f27c4075e6b0fe7db11c8a3)).
+
+### 🧪 Patch [0.0.2](https://github.com/gabrielba15/portfolio/releases/tag/v0.0.2) (Dec 30, 2021)
+
+#### Features
+
+- Alt Name Gaboland ([b0a8ebc0c](https://github.com/gabrielba15/portfolio/commit/b0a8ebc0cc38579c68b22a967ba814556af1b3aa)).
+
+### 🧪 Patch [0.0.1](https://github.com/gabrielba15/portfolio/releases/tag/v0.0.1) (Dec 29, 2021)
+
+#### Features
+
+- Initial Commit ([173d18f8f](https://github.com/gabrielba15/portfolio/commit/173d18f8fbd7862a7e1ee63a9723fb5d2c438d29)).